% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_functions.R
\name{download_video}
\alias{download_video}
\title{Download the video}
\usage{
download_video(id, api_url, output_path)
}
\arguments{
\item{id}{the ID returned from running generate_from_gs}

\item{output_path}{a file path to download the video to}

\item{...}{Other parameters passed to \code{httr::GET()}}
}
\value{
Download video to
}
\description{
Download the video
}
