% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_functions.R
\name{generate_from_gs}
\alias{generate_from_gs}
\title{Generate video from Google Slides}
\usage{
generate_from_gs(
  link,
  service = "coqui",
  model_name = "jenny",
  vocoder_name = "jenny",
  api_url = api_url(),
  ...
)
}
\arguments{
\item{link}{Full URL to Google Slides with public access enabled}

\item{service}{Speech synthesis service}

\item{model_name}{The voice used to synthesize the audio}

\item{vocoder_name}{Voice coder used for speech coding and transmission}

\item{api_url}{URL that contains API}

\item{...}{Other parameters passed to \code{httr::GET()}}
}
\value{
Response from the API
}
\description{
Generate video from Google Slides
}
