% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_functions.R
\name{mario_health_check}
\alias{mario_health_check}
\title{Health Check}
\usage{
mario_health_check(api_url = mario_api_url())
}
\arguments{
\item{api_url}{URL for API}
}
\description{
Check if the API is running
}
