% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_functions.R
\name{post_video_creation}
\alias{post_video_creation}
\title{Generate video from Google Slides}
\usage{
post_video_creation(
  link,
  service = "coqui",
  model_name = "jenny",
  vocoder_name = "jenny",
  api_url = loqui_api_url(),
  ...
)
}
\arguments{
\item{link}{Full URL to Google Slides with public access enabled}

\item{service}{Speech synthesis service}

\item{model_name}{The voice used to synthesize the audio}

\item{vocoder_name}{Voice coder used for speech coding and transmission}

\item{api_url}{URL that contains API}

\item{...}{Other parameters passed to \code{httr::GET()}}

\item{output}{Path video output}
}
\value{
Path to video output
}
\description{
Generate video from Google Slides
}
