% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_functions.R
\name{request_status}
\alias{request_status}
\title{Check on status of video POST request}
\usage{
request_status(id, api_url)
}
\arguments{
\item{id}{the ID returned from running generate_from_gs}

\item{...}{Other parameters passed to \code{httr::GET()}}
}
\value{
Path to video output
}
\description{
Check on status of video POST request
}
